Rename org.jdbi.v3.stringsubstitor package to .commonstext
